Gør random number til float

+++ Util/LFO/lfo.hpp
@@ -91,7 +91,7 @@ enum lfo_mode{sine,square,triangle,sampleHold};
 				// Triggers at waveShape-level and at 0 ('ish)
 				if (isPhaseUnderWaveshapeLevel() != waveShapeStatusState)
 				{
-					LFO_value = ((double)rand() / (1));
+					LFO_value = static_cast <float> (rand()) / static_cast <float> (1);
 				}
 				//toggle waveShape status state
 				waveShapeStatusState = !waveShapeStatusState;
